split fn; datasetutils passing
